requirements.txt
behave==1.2.6
pytest==8.3.5
requests==2.32.3
selenium==4.30.0
webdriver-manager==4.0.2
Appium-Python-Client==5.0.0